+++ commander/executor/create_create2transfer_commitments.go
@@ -25,7 +25,7 @@ func (t *TransactionExecutor) CreateCreate2TransferCommitments(domain *bls.Domai
 	}
 	commitments = make([]models.Commitment, 0, t.cfg.MaxCommitmentsPerBatch)
 
-	for i := uint32(0); len(commitments) != int(t.cfg.MaxCommitmentsPerBatch); i++ {
+	for i := uint8(0); len(commitments) != int(t.cfg.MaxCommitmentsPerBatch); i++ {
 		var commitment *models.Commitment
 		commitmentID.IndexInBatch = i
