+++ models/transaction_base.go
@@ -16,7 +16,7 @@ type TransactionBase struct {
 	ReceiveTime  *Timestamp `db:"receive_time"`
 	ErrorMessage *string    `db:"error_message"`
 	BatchID      *Uint256   `db:"batch_id"` //TODO: use CommitmentID struct with badger
-	IndexInBatch *uint32    `db:"index_in_batch"`
+	IndexInBatch *uint8     `db:"index_in_batch"`
 }
 
 type TransactionBaseForCommitment struct {
